feat: implement pg-data-connector
Postgres Data Connector use JDBC Sink Connector to streaming data from
specific Kafka topic to Postgres database table.

This commit implement all the necessary configuration and changes.

Currently there is a bug that caused Postgres Data Connector die.
This is due to the unique postgres constraint on the table columns it's writing
to. This issue will be resolve in the next commit.