fix: record deletion doesn't works for Postgres CDC Connector
Delection of keyword record in Postgres will cause Postgres CDC
Connector failed. This is due to My wrong understanding with Debezium Postgres
Connector.

This commit fix that. the code look bad because it work around many bad
database schema design decision. I shouldn't insist to add a id record
as a primary key to everything when there is a better option.