fix: better approach to update schema
The schema-updater is used to upload avro schema definition to
schema-registry.

Registry might take times to startup, if the schema-updater
launch too early. The updater will failed.

The old way we remedy this is to let updater wait for a while before
sending request, But it didn't solve this issue very well.

Now we use docker-compose restart-policy to handle this. Probably
better.